Ajustes de pacotes e nomenclaturas
